upload version 3.3
Made the 'my_net_id_screen' more intuitive, and it shoudl restrict bad input and things like that. still need to format that keypad so all keys make it on the screen for smaller screens like the iphone 8, right now the done and del buttons are hidden unless you scroll the keyboard down, which is rediculous, obviously.