added the latest build
this build archive was actually produced yesterday on 7/9/2024. The difference in the build acquired from  running this versus the last one that was on here is location screen and how it is just a pop up now. we still need to test whether this actually made it so that the notifications still work when we have the locatin screen open. Actually, I think it would work no matter what for any screens since they are a child and technically running over the top of the device screen which should still be running.